+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Important changes will be notified in this file
 
 ## unreleased
 - Fixed: do not record order update events when state hasn't changed.
+- Added: ProductDetail::getUnitPriceAsPrice() and ProductDetail::getSalePriceAsPrice() to retrieve the original Price objects.
 
 ## 2022-11-07 - 0.5.2
 - Fixed: localized variant option title returns full array when locale title wasn't present